cbor_smol::ser

Trait Writer

pub trait Writer {
    type Error: Into<Error>;

    // Required method
    fn write_all(&mut self, buf: &[u8]) -> Result<(), Self::Error>;
}

Required Associated Types

type Error: Into<Error>

The type of error returned when a write operation fails.

Required Methods

fn write_all(&mut self, buf: &[u8]) -> Result<(), Self::Error>

Attempts to write an entire buffer into this write.