Trait cbor_smol::ser::Writer

pub trait Writer {
    type Error: Into<Error>;
    fn write_all(&mut self, buf: &[u8]) -> Result<(), Self::Error>;
}

Associated Types

type Error: Into<Error>

The type of error returned when a write operation fails.

Loading content...
Required methods

fn write_all(&mut self, buf: &[u8]) -> Result<(), Self::Error>

Attempts to write an entire buffer into this write.

Loading content...